+++ packages/ecount-dev-tool/eslint.config.ts
@@ -1,7 +1,4 @@
 import { createSvelteConfig } from '../../eslint.config.ts';
 import globals from 'globals';
 
-export default [
-    ...createSvelteConfig({ ...globals.webextensions }, import.meta.dirname),
-    { ignores: ['src/old/**'] },
-];
+export default [...createSvelteConfig({ ...globals.webextensions }, import.meta.dirname), { ignores: ['src/old/**'] }];